requirements.txt
Flask==2.*
gunicorn==20.*
psycopg2-binary==2.9.*
